build(depracated): update `zarr` due to a security issue.
FULL EXPLANATION (for future reference!):

I tried to `uv add` a new package, but got:

```
warning: `zarr==3.0.3` is yanked (reason: "broke a certain indexing incantation with our last release. see #2849. This can lead to data corruption")
```
(The issue being referenced: https://github.com/zarr-developers/zarr-python/issues/2849 )

For reference, "yanking" a package means marking a specific version as withdrawn or deprecated due to a serious issue. The version is still available for projects that already depend on it, bkut it's hidden from new dependency resolution so that it won't be chosen by default.

In other words, the issue is with `zarr`, not `uv`. To upgrade:

`uv sync --all-extras --all-groups --upgrade-package zarr`