refactor(GRS): change data source and corresponding loading method...
I previously got data from <https://repository.lsu.edu/geo_psl/1/> which provides multiple separate text files -- you don't need external libraries to read the data, but I had to write a custom script that zips them together (which is a bit hacky/overcomplicated because you need to change the "last modified" datetime to ensure a consistent hash).

In the original paper <https://doi.org/10.1029/2022GL099235>, turns out they also uploaded the data here <https://data.mendeley.com/datasets/3jd9whd78m/1> as a single XLSX, also with a DOI! You need an additional dependency to read the XLSX with pandas (`python-calamine`) but it's not that big (~1-3 MB?). So we use that instead to get rid of the complexity of `generate.py` and unzipping in RAM.

Only potentially breaking change is whether sum of volatiles ('cl+h2o+s') in undefined regions is nan or 0.